refactor: remove Settings component and related logic, streamline organization settings route
